MapPanel now renders SimpleRoad instances
I introduced the concept of waypoints for the purpose of drawing roads
and for calculating intersections. The SimpleRoad now knows how to
calculate its own waypoints, and these get used by the MapPanel to
render the SimpleRoad.

TimedCoordinates have also been added, to give a sense of time scale
between waypoints. This will be used to allow cars to travel at the
appropriate speeds along the roads in the GUI, and can potentially be
used if we want to animate the painting of the roads.