Painting roads between intersections
Not yet painting sources/sinks. Roads are currently drawn without taking
the road length into account (they simply connect the intersection
points), but we can express the car's position as a ratio along the road
and scale appropriately. The only discrepancy between painting the
actual length of the road (200 units) vs connecting the intersection
points should be the size of the intersection itself, but I'm not sure
how that really plays into the code yet.