Muahahaha
This commit has some nice cosmetic changes, but the more important parts
are under the hood. Driver still gets things started, but I created a
Simulation object to encapsulate everything with a single simulation run.
We will have to run many generations with many individuals in each
generations, so this is a step on the way to doing that.

Also, the SimulationPanel now owns the Simulation, which is nice because
a Simulation can exist outside any UI components (i.e. headless).

I also removed some of the references to Driver in other classes to
reduce coupling, but there are still 4 references to Driver.frameRate I
believe. I would like to get rid of them, but I would need to figure out
how to do that.

Also, the SwingWorker stuff runs on another thread, so I think we should
be able to get this running on all cylinders soon.

Bonus: I think the code is now also compatible with previous version of Java.

Next steps:
- code to import/export results to/from a file. Importing might help if we
want to start running, then stop and resume later, potentially on other
hardware.
- UI and multithreading code (multithreading will be trivial with
SwingWorkers) to start running these in batch mode and in parallel.
Visualizing progress of each run as it goes would definitely be
preferable, but not necessary as long as we know how many have finished.