Starting to implement genetic algorithm steps
This is a bit of a work in progress, hence the nasty printlns
everywhere. "Survival of the fittest," crossover, mutations, and
cloning are theoretically implemented. However, the 2nd cloned
individual was definitely not correct, so there's some silly bug to fix
that I think is breaking all the other chromosomes as well (since the
ones theoretically derived from parents are not even close to either).

However, I think I'm really close to having everything working on the
command line. I'd eventually like a GUI to make everything cleaner
though. I've started working on it, but put it on the backburner until
the backend appears to be fully functional.

I also added a Results object that comprises the results from a
Simulation run. It works nicely with the java.util.concurrent.Future
objects I'm now using for the parallelism. Those are nice because I can
return results from my threads, unlike runnable's run() method.